login 브랜치 기능 적용 확인(토큰 작업 필요) 및 Admin-ProductList 부분 작성 필요

+++ src/main/java/web/ssa/entity/Payment.java
@@ -7,14 +7,15 @@ import java.time.LocalDateTime;
 @Entity
 @Data
 public class Payment {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Long productId;
+    private int productId;
     private String itemName;
     private int amount;
     private String status; // SUCCESS / REFUND_REQUEST / REFUNDED / CANCELLED
-    private String tid;    // 카카오 결제 TID
+    private String tid; // 카카오 결제 TID
     private String userEmail;
 
     private LocalDateTime createdAt = LocalDateTime.now();
